[testing] move fixtures to main library
So that clients can encode valid tokens in tests.
